+++ tests/epyccel/test_epyccel_lists.py
@@ -536,6 +536,24 @@ def test_extend_list_class_attribute(language):
     accelerated_list = modnew.fn()
     assert python_list == accelerated_list
 
+def test_copy_basic(language):
+    def f():
+        a = [1, 2, 3]
+        b = a.copy()
+        return b
+
+    epyc_f = epyccel(f, language=language)
+    assert f() == epyc_f()
+
+def test_copy_nested(language):
+    def f():
+        a = [[1, 2], [3, 4]]
+        b = a.copy()
+        return b
+
+    epyc_f = epyccel(f, language=language)
+    assert f() == epyc_f()
+
 def test_mixed_list_methods(language):
     def f():
         a = [(1, 4, 5), (33, 12, 5), (3, 5)]